+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2 (beta)
 
 - Added support for `$BROWSER` environment variable on CI.
+- Added `fake.person.honorific` property.
 
 # 0.1 (2020-07-03)